LayoutTests/http/tests/navigation/page-cache-xhr-expected.txt

Tests that a page with a loading XMLHttpRequest goes into the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
PASS Executed the XHR error handler before entering PageCache
PASS xhr.status is 0
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS successfullyParsed is true

TEST COMPLETE
